Create seperate files for Result and UpdateInfoResponse
